Move more aws service interactions to a separate package
